refactoring

+++ controllers/Character.js
@@ -1,7 +1,4 @@
-const axios = require("axios");
-const event = require('../events/createFile');
 const Character = require('../models/Character');
-const BLIZZARD_URL = "https://backend-tp-final-nodejs.agence-pixi.fr/wow/compte/check";
 
 const createCharacter = (req, res, next) => {    
     delete req.body._id;
@@ -12,9 +9,9 @@ const createCharacter = (req, res, next) => {
         class: req.body.class
     })
         .then((character) => {
-            if (character) return res.status(401).json({ message: 'Pseudo/classe déjà utilisés' });
-
             if (req.auth.isAdmin) return res.status(401).json({ message: "Création refusée en tant qu'administrateur" })
+
+            if (character) return res.status(401).json({ message: 'Pseudo/classe déjà utilisés' });
 
             const characterObject = new Character({
                 ...req.body,
@@ -29,40 +26,36 @@ const createCharacter = (req, res, next) => {
         .catch((error) => res.status(400).json({ error }));
 }
 
-const blizzardCreateCharacter = (req, res, next) => {
-    axios.post(BLIZZARD_URL, { "username": req.body.username, "password": req.body.password })
-        .then((response) => res.status(201).json(response.data))
-        .catch((error) => {
-            event.emit('createFile', { file: 'blizzardCreateLogs', message: `${new Date().toLocaleDateString()} - ${new Date().toLocaleTimeString()} : Tentative de connexion invalide\n` });
-
-            res.status(400).json({ error })
-        });
-}
-
 const getCharacter = (req, res, next) => {
     Character.findOne({
         pseudo: req.params.pseudo,
         class: req.params.class
     })
         .then((character) => {
-            if (!character) res.status(401).json({ message: 'Personnage inexistant' });
+            if (!character) return res.status(401).json({ message: 'Personnage inexistant' });
 
             res.status(200).json(character);
         })
         .catch((error) => res.status(400).json({ error }));
 }
 
-const getCharacters = (req, res, next) => {
+const getCharacterById = (req, res, next) => {
     Character.findById(req.params.id)
-        .then((character) => res.status(200).json(character))
+        .then((character) => {
+            if (!character) return res.status(401).json({ message: 'Identifiant inexistant' });
+
+            res.status(200).json(character);
+        })
         .catch((error) => res.status(400).json(error));
 }
 
 const updateCharacter = (req, res, next) => {
     delete req.body.userId;
 
-    Character.findOne({ _id: req.params.id })
+    Character.findById(req.params.id)
     .then((character) => {
+        if (!character) return res.status(401).json({ message: 'Identifiant inexistant' });
+
         if (character.userId === req.auth.userId || req.auth.isAdmin) {
             Character.updateOne({ _id: req.params.id }, { ...req.body, _id: req.params.id })
                 .then(() => res.status(200).json({ message: 'Personnage modifié' }))
@@ -71,12 +64,14 @@ const updateCharacter = (req, res, next) => {
             res.status(401).json({ message: 'Cet utilisateur ne peut pas modifier ce personnage' })
         }
     })
-    .catch((error) => res.status(400).json({ error }));
+    .catch(() => res.status(400).json({ error }));
 }
 
 const deleteCharacter = (req, res, next) => {
     Character.findById(req.params.id)
         .then((character) => {
+            if (!character) return res.status(401).json({ message: 'Identifiant inexistant' });
+
             if (character.userId === req.auth.userId || req.auth.isAdmin) {
                 Character.deleteOne({ _id: req.params.id})
                     .then(() => res.status(200).json({ message: 'Personnage supprimé' }))
@@ -85,14 +80,13 @@ const deleteCharacter = (req, res, next) => {
                 res.status(401).json({ message: 'Cet utilisateur ne peut pas supprimer ce personnage' });
             }
         })
-        .catch((error) => res.status(400).json({ error }));
+        .catch(() => res.status(400).json({ error }));
 }
 
 module.exports = {
     createCharacter,
-    blizzardCreateCharacter,
     getCharacter,
-    getCharacters,
+    getCharacterById,
     updateCharacter,
     deleteCharacter
 }
